+++ core/vm/actor/builtin/v0/market/market_actor_utils.cpp
@@ -11,7 +11,7 @@
 #include "vm/actor/builtin/types/market/policy.hpp"
 #include "vm/actor/builtin/v0/miner/miner_actor.hpp"
 #include "vm/actor/builtin/v0/reward/reward_actor.hpp"
-#include "vm/actor/builtin/v0/storage_power/storage_power_actor_export.hpp"
+#include "vm/actor/builtin/v0/storage_power/storage_power_actor.hpp"
 #include "vm/actor/builtin/v0/verified_registry/verified_registry_actor.hpp"
 #include "vm/toolchain/toolchain.hpp"
