+++ core/vm/actor/builtin/v0/storage_power/CMakeLists.txt
@@ -4,7 +4,7 @@
 #
 
 add_library(storage_power_actor_v0
-    storage_power_actor_export.cpp
+    storage_power_actor.cpp
     storage_power_actor_utils.cpp
     )
 target_link_libraries(storage_power_actor_v0
