+++ core/vm/actor/builtin/v3/miner/miner_actor.cpp
@@ -7,7 +7,7 @@
 
 #include "vm/actor/builtin/states/miner/miner_actor_state.hpp"
 #include "vm/actor/builtin/v3/account/account_actor.hpp"
-#include "vm/actor/builtin/v3/storage_power/storage_power_actor_export.hpp"
+#include "vm/actor/builtin/v3/storage_power/storage_power_actor.hpp"
 #include "vm/toolchain/toolchain.hpp"
 
 namespace fc::vm::actor::builtin::v3::miner {
