+++ core/vm/actor/builtin/v3/storage_power/CMakeLists.txt
@@ -4,7 +4,7 @@
 #
 
 add_library(storage_power_actor_v3
-    storage_power_actor_export.cpp
+    storage_power_actor.cpp
     )
 target_link_libraries(storage_power_actor_v3
     storage_power_actor_state
