+++ core/vm/actor/impl/invoker_impl.cpp
@@ -13,7 +13,7 @@
 #include "vm/actor/builtin/v0/multisig/multisig_actor.hpp"
 #include "vm/actor/builtin/v0/payment_channel/payment_channel_actor.hpp"
 #include "vm/actor/builtin/v0/reward/reward_actor.hpp"
-#include "vm/actor/builtin/v0/storage_power/storage_power_actor_export.hpp"
+#include "vm/actor/builtin/v0/storage_power/storage_power_actor.hpp"
 #include "vm/actor/builtin/v0/system/system_actor.hpp"
 #include "vm/actor/builtin/v0/verified_registry/verified_registry_actor.hpp"
 
@@ -25,7 +25,7 @@
 #include "vm/actor/builtin/v2/multisig/multisig_actor.hpp"
 #include "vm/actor/builtin/v2/payment_channel/payment_channel_actor.hpp"
 #include "vm/actor/builtin/v2/reward/reward_actor.hpp"
-#include "vm/actor/builtin/v2/storage_power/storage_power_actor_export.hpp"
+#include "vm/actor/builtin/v2/storage_power/storage_power_actor.hpp"
 #include "vm/actor/builtin/v2/system/system_actor.hpp"
 #include "vm/actor/builtin/v2/verified_registry/verified_registry_actor.hpp"
 #include "vm/actor/cgo/actors.hpp"
