+++ test/core/vm/actor/builtin/v0/miner/miner_actor_v0_test.cpp
@@ -15,7 +15,7 @@
 #include "vm/actor/builtin/types/miner/v0/deadline.hpp"
 #include "vm/actor/builtin/states/miner/v0/miner_actor_state.hpp"
 #include "vm/actor/builtin/v0/miner/miner_actor_utils.hpp"
-#include "vm/actor/builtin/v0/storage_power/storage_power_actor_export.hpp"
+#include "vm/actor/builtin/v0/storage_power/storage_power_actor.hpp"
 #include "vm/actor/codes.hpp"
 
 namespace fc::vm::actor::builtin::v0::miner {
