+++ test/core/vm/actor/builtin/v0/power/storage_power_actor_v0_test.cpp
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "vm/actor/builtin/v0/storage_power/storage_power_actor_export.hpp"
+#include "vm/actor/builtin/v0/storage_power/storage_power_actor.hpp"
 #include "vm/actor/builtin/states/storage_power/v0/storage_power_actor_state.hpp"
 
 #include <gtest/gtest.h>
