+++ test/core/vm/actor/builtin/v2/miner/miner_actor_v2_test.cpp
@@ -14,7 +14,7 @@
 #include "vm/actor/builtin/types/miner/policy.hpp"
 #include "vm/actor/builtin/types/miner/v2/deadline.hpp"
 #include "vm/actor/builtin/states/miner/v2/miner_actor_state.hpp"
-#include "vm/actor/builtin/v2/storage_power/storage_power_actor_export.hpp"
+#include "vm/actor/builtin/v2/storage_power/storage_power_actor.hpp"
 #include "vm/actor/codes.hpp"
 
 namespace fc::vm::actor::builtin::v2::miner {
